Favicons and Device Icons Added
Upon noticing that there wasn't a favicon, I looked into the code and noticed they were disabled. I reached out to Ben Centra and discovered they were just disabled because he couldn't get them working. Not for any reason of breaking code. I generated a series of Icons for the Material logo and set them up.
As well as that I fixed an slight error in the README for the LDAP Config and added jetbrains .idea to the gitignore.
Fixes #47